loader: Move compile code to compile.go
This code is related to compilation, move it to compile.go.
